filesystem tweaks to compile w/C++20
